# app/swagger.yaml
swagger: "2.0"
info:
  title: "Banking API"
  description: "API documentation for the Banking application."
  version: "1.0.0"
host: "localhost:5000"
basePath: "/api"
schemes:
  - "http"
...
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: "Enter your JWT token in the format **Bearer &lt;token&gt;**"
paths:
  /auth/register:
    post:
      summary: "Register a new user"
      description: "Endpoint to register a new user with email, password, and optional first/last name."
      tags:
        - "Authentication"
      parameters:
        - in: "body"
          name: "body"
          required: true
          description: "User registration details"
          schema:
            type: "object"
            required:
              - email
              - password
            properties:
              email:
                type: "string"
                format: "email"
                example: "user@example.com"
              password:
                type: "string"
                format: "password"
                example: "SecurePassword123!"
              username:
                type: "string"
                example: "user123"
              first_name:
                type: "string"
                example: "John"
              last_name:
                type: "string"
                example: "Doe"
      responses:
        201:
          description: "User registered successfully"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "User registered successfully"
              user:
                type: "object"
                properties:
                  id:
                    type: "integer"
                    example: 1
                  username:
                    type: "string"
                    example: "user123"
                  email:
                    type: "string"
                    example: "user@example.com"
                  first_name:
                    type: "string"
                    example: "John"
                  last_name:
                    type: "string"
                    example: "Doe"
        400:
          description: "Validation error"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Missing required fields"
        409:
          description: "Conflict error"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Email already exists"
  /auth/login:
    post:
      summary: "User Login"
      description: "Authenticate a user and return access and refresh tokens."
      tags:
        - "Authentication"
      parameters:
        - in: "body"
          name: "body"
          required: true
          description: "User login credentials"
          schema:
            type: "object"
            required:
              - password
            properties:
              email:
                type: "string"
                format: "email"
                example: "user@example.com"
              username:
                type: "string"
                example: "user123"
              password:
                type: "string"
                format: "password"
                example: "SecurePassword123!"
      responses:
        200:
          description: "Login successful"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Login successful"
              token:
                type: "string"
                example: "access_token_here"
              refresh_token:
                type: "string"
                example: "refresh_token_here"
        401:
          description: "Invalid credentials"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Invalid credentials"
...
  /auth/refresh:
    post:
      summary: "Refresh Token"
      description: "Endpoint to refresh token using refresh token in request header."
      tags:
        - "Authentication"
      security:
        - Bearer: []
      responses:
        200:
          description: "Token refreshed successfully"
          schema:
            type: "object"
            properties:
              token:
                type: "string"
                example: "new_access_token_here"
              refresh_token:
                type: "string"
                example: "new_refresh_token_here"
        401:
          description: "Invalid or expired refresh token"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Invalid token identity"
  /auth/logout:
    post:
      summary: "Logout User"
      description: "Endpoint to log out user by revoking their JWT token."
      tags:
        - "Authentication"
      security:
        - Bearer: []
      responses:
        200:
          description: "Successfully logged out"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Successfully logged out"
  /auth/profile:
    get:
      summary: "Get User Profile"
      description: "Retrieve the authenticated user's profile."
      tags:
        - "Authentication"
      security:
        - Bearer: []
      responses:
        200:
          description: "User profile retrieved successfully"
          schema:
            type: "object"
            properties:
              id:
                type: "integer"
                example: 1
              username:
                type: "string"
                example: "user123"
              email:
                type: "string"
                example: "user@example.com"
              first_name:
                type: "string"
                example: "John"
              last_name:
                type: "string"
                example: "Doe"
        404:
          description: "User not found"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "User not found"
  /auth/verify:
    post:
      summary: "Verify Token"
      description: "Verify if a token is valid and not expired."
      tags:
        - "Authentication"
      security:
        - Bearer: []
      responses:
        200:
          description: "Token is valid"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Token is valid"
              verified:
                type: "boolean"
                example: true
  /auth/change-password:
    post:
      summary: "Change Password"
      description: "Endpoint to change a user's password."
      tags:
        - "Authentication"
      security:
        - Bearer: []
      parameters:
        - in: "body"
          name: "body"
          required: true
          description: "Password change details"
          schema:
            type: "object"
            required:
              - current_password
              - new_password
            properties:
              current_password:
                type: "string"
                format: "password"
                example: "OldPassword123!"
              new_password:
                type: "string"
                format: "password"
                example: "NewSecurePassword123!"
      responses:
        200:
          description: "Password changed successfully"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Password changed successfully"
        401:
          description: "Current password is incorrect"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Current password is incorrect"
        400:
          description: "New password does not meet complexity requirements"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "New password must meet complexity requirements"
...